Start server with flask script
