+++ README.md
@@ -1,0 +1,6 @@
+# pipecontrol
+Website for interactions with our two-photon pipeline.
+
+# Known issues
+* Run with docker-compose up (not docker-compose run).
+* Make sure startserver.sh is executable before building the container. Throws 'Permission denied' error otherwise.
