--- app/main/__init__.py
@@ -1,6 +1,0 @@
-from flask import Blueprint
-
-main = Blueprint('main', __name__)
-
-from . import views, errors
-
