--- app/main/decorators.py
@@ -1,10 +1,0 @@
-
-import datajoint as dj
-
-def ping(f):
-    def ret(*args, **kwargs):
-        dj.conn()
-        return f(*args, **kwargs)
-    return ret
-
-
