--- app/main/utils.py
@@ -1,6 +1,0 @@
-import random
-import string
-
-def namehash(N=20):
-    return ''.join(random.choice(string.ascii_uppercase + string.digits) for _ in range(N))
-
